Update fdb-relational-api gradle so that it builds

+++ fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/exceptions/InternalErrorException.java
@@ -20,8 +20,12 @@
 
 package com.apple.foundationdb.relational.api.exceptions;
 
+/**
+ * Internal type of relational exception. Considered for removal to replace with using
+ * {@link RelationalException} directly.
+ */
 public class InternalErrorException extends RelationalException {
-    private static final int serialVersionUID = 1;
+    private static final long serialVersionUID = 1;
 
     public InternalErrorException(String message) {
         super(message, ErrorCode.INTERNAL_ERROR);
